Add Center option for File in file-text layout.
The Fugitive Slave exhibition uses a new CUL layout created just before
the upgrade to Omeka 2.2.2 and Exhibit Builder 3.1.3. To move the
exhibiton to the new version of Exhibit Builder, a Center option was added
to have a large image centered on the page, as used in the CUL layout.
Once a new block layout for file only is created by CUL, this
Center option will be removed because it does not belong in
file-text layout, since there should be no text in a block with a large
image  -- text should be added in a text layout block before or after
the file only block.