Remove option file-position:center for file-text layout
The above option was added for the "Record of Fugitives" exhibit during
the upgrade to Omeka 2.2.2 and Exhibit Builder 3.1.3 to support the
display of a centered, full-width, image. It was added to the file-text
layout, which is one of the layouts packaged with Exhibit Builder.
However, it was always planned that this customization would be
moved to its own layout, and the file-text layout would revert back
to its original state, without a file-position:center option.
The new layout has be created (see CulLayouts plugin) and the
"Record of Fugitives" exhibit has been updated to use the new layout.
So now we can remove the customization from the file-text layout.